Create XCDBLD.framework
