+++ Source/CarthageKit/Availability.swift
@@ -5,6 +5,7 @@ import ReactiveSwift
 import ReactiveCocoa
 #endif
 import Result
+import XCDBLD
 
 // MARK: - Archive.swift
