+++ Source/CarthageKit/BuildSettings.swift
@@ -5,6 +5,7 @@ import ReactiveSwift
 #else
 import ReactiveCocoa
 #endif
+import XCDBLD
 
 /// A map of build settings and their values, as generated by Xcode.
 public struct BuildSettings {
@@ -12,9 +13,9 @@ public struct BuildSettings {
 	public let target: String
 
 	/// All build settings given at initialization.
-	public let settings: Dictionary<String, String>
+	public let settings: [String: String]
 
-	public init(target: String, settings: Dictionary<String, String>) {
+	public init(target: String, settings: [String: String]) {
 		self.target = target
 		self.settings = settings
 	}
