+++ Source/CarthageKit/XCDBLDExtensions.swift
@@ -1,37 +1,25 @@
-import Foundation
-#if swift(>=3)
 import ReactiveSwift
-#else
-import ReactiveCocoa
-#endif
+import ReactiveTask
+import Result
+import XCDBLD
 
-/// Describes how to locate the actual project or workspace that Xcode should
-/// build.
-public enum ProjectLocator: Comparable {
-	/// The `xcworkspace` at the given file URL should be built.
-	case workspace(URL)
-
-	/// The `xcodeproj` at the given file URL should be built.
-	case projectFile(URL)
-
-	/// The file URL this locator refers to.
-	public var fileURL: URL {
-		switch self {
-		case let .workspace(url):
-			assert(url.isFileURL)
-			return url
-
-		case let .projectFile(url):
-			assert(url.isFileURL)
-			return url
-		}
+extension MachOType {
+	/// Attempts to parse a Mach-O type from a string returned from `xcodebuild`.
+	public static func from(string: String) -> Result<MachOType, CarthageError> {
+		return Result(self.init(rawValue: string), failWith: .parseError(description: "unexpected Mach-O type \"\(string)\""))
 	}
+}
 
-	/// The number of levels deep the current object is in the directory hierarchy.
-	public var level: Int {
-		return fileURL.carthage_pathComponents.count - 1
+extension Platform {
+	/// The relative path at which binaries corresponding to this platform will
+	/// be stored.
+	public var relativePath: String {
+		let subfolderName = rawValue
+		return (CarthageBinariesFolderPath as NSString).appendingPathComponent(subfolderName)
 	}
+}
 
+extension ProjectLocator {
 	/// Attempts to locate projects and workspaces within the given directory.
 	///
 	/// Sends all matches in preferential order.
@@ -65,7 +53,7 @@ public enum ProjectLocator: Comparable {
 			.map { $0.sorted() }
 			.flatMap(.merge) { SignalProducer<ProjectLocator, CarthageError>($0) }
 	}
-
+	
 	/// Sends each scheme found in the receiver.
 	public func schemes() -> SignalProducer<String, CarthageError> {
 		let task = xcodebuildTask("-list", BuildArguments(project: self))
@@ -76,7 +64,7 @@ public enum ProjectLocator: Comparable {
 			// xcodebuild has a bug where xcodebuild -list can sometimes hang
 			// indefinitely on projects that don't share any schemes, so
 			// automatically bail out if it looks like that's happening.
-			.timeout(after: 60, raising: .xcodebuildTimeout(self), on: QueueScheduler(qos: .default))
+			.timeout(after: 60, raising: .xcodebuildTimeout(self), on: QueueScheduler())
 			.retry(upTo: 2)
 			.map { data in
 				return String(data: data, encoding: .utf8)!
@@ -102,42 +90,17 @@ public enum ProjectLocator: Comparable {
 	}
 }
 
-public func ==(_ lhs: ProjectLocator, _ rhs: ProjectLocator) -> Bool {
-	switch (lhs, rhs) {
-	case let (.workspace(left), .workspace(right)):
-		return left == right
-
-	case let (.projectFile(left), .projectFile(right)):
-		return left == right
-
-	default:
-		return false
+extension SDK {	
+	/// Attempts to parse an SDK name from a string returned from `xcodebuild`.
+	public static func from(string: String) -> Result<SDK, CarthageError> {
+		return Result(self.init(rawValue: string.lowercased()), failWith: .parseError(description: "unexpected SDK key \"\(string)\""))
 	}
-}
-
-public func <(_ lhs: ProjectLocator, _ rhs: ProjectLocator) -> Bool {
-	// Prefer top-level directories
-	let leftLevel = lhs.level
-	let rightLevel = rhs.level
-	guard leftLevel == rightLevel else {
-		return leftLevel < rightLevel
-	}
-
-	// Prefer workspaces over projects.
-	switch (lhs, rhs) {
-	case (.workspace, .projectFile):
-		return true
-
-	case (.projectFile, .workspace):
-		return false
-
-	default:
-		return lhs.fileURL.carthage_path.characters.lexicographicallyPrecedes(rhs.fileURL.carthage_path.characters)
-	}
-}
-
-extension ProjectLocator: CustomStringConvertible {
-	public var description: String {
-		return fileURL.carthage_lastPathComponent
+	
+	/// Split the given SDKs into simulator ones and device ones.
+	internal static func splitSDKs<S: Sequence>(_ sdks: S) -> (simulators: [SDK], devices: [SDK]) where S.Iterator.Element == SDK {
+		return (
+			simulators: sdks.filter { $0.isSimulator },
+			devices: sdks.filter { !$0.isSimulator }
+		)
 	}
 }
