+++ Source/XCDBLD/XCDBLD.h
@@ -1,0 +1,16 @@
+//
+//  XCDBLD.h
+//  XCDBLD
+//
+//  Created by Matt Diephouse on 2017-01-23.
+//  Copyright (c) 2014 Carthage. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for XCDBLD.
+FOUNDATION_EXPORT double XCDBLDVersionNumber;
+
+//! Project version string for CarthageKit.
+FOUNDATION_EXPORT const unsigned char XCDBLDVersionString[];
+
