+++ Source/XCDBLDTests/ProjectLocatorSpec.swift
@@ -1,4 +1,4 @@
-import CarthageKit
+import XCDBLD
 import Foundation
 import Nimble
 import Quick
