+++ XCDBLD.xcodeproj/xcshareddata/xcschemes/XCDBLD.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "0710"
+   LastUpgradeVersion = "0820"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "D0E7B63119E9C64500EDBA4D"
-               BuildableName = "carthage.app"
-               BlueprintName = "carthage"
-               ReferencedContainer = "container:Carthage.xcodeproj">
+               BlueprintIdentifier = "D0D1216C19E87B05005E4BAA"
+               BuildableName = "XCDBLD.framework"
+               BlueprintName = "XCDBLD"
+               ReferencedContainer = "container:XCDBLD.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -38,86 +38,56 @@
                ReferencedContainer = "container:XCDBLD.xcodeproj">
             </BuildableReference>
          </TestableReference>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "D0D1217619E87B05005E4BAA"
-               BuildableName = "CarthageKitTests.xctest"
-               BlueprintName = "CarthageKitTests"
-               ReferencedContainer = "container:Carthage.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
       </Testables>
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "D0E7B63119E9C64500EDBA4D"
-            BuildableName = "carthage.app"
-            BlueprintName = "carthage"
-            ReferencedContainer = "container:Carthage.xcodeproj">
+            BlueprintIdentifier = "D0D1216C19E87B05005E4BAA"
+            BuildableName = "XCDBLD.framework"
+            BlueprintName = "XCDBLD"
+            ReferencedContainer = "container:XCDBLD.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </TestAction>
    <LaunchAction
-      buildConfiguration = "Release"
+      buildConfiguration = "Debug"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       launchStyle = "0"
       useCustomWorkingDirectory = "NO"
       ignoresPersistentStateOnLaunch = "NO"
-      debugDocumentVersioning = "NO"
+      debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "D0E7B63119E9C64500EDBA4D"
-            BuildableName = "carthage.app"
-            BlueprintName = "carthage"
-            ReferencedContainer = "container:Carthage.xcodeproj">
+            BlueprintIdentifier = "D0D1216C19E87B05005E4BAA"
+            BuildableName = "XCDBLD.framework"
+            BlueprintName = "XCDBLD"
+            ReferencedContainer = "container:XCDBLD.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
-      <CommandLineArguments>
-         <CommandLineArgument
-            argument = "update"
-            isEnabled = "YES">
-         </CommandLineArgument>
-      </CommandLineArguments>
-      <EnvironmentVariables>
-         <EnvironmentVariable
-            key = "DEVELOPER_DIR"
-            value = "${DEVELOPER_DIR}"
-            isEnabled = "YES">
-         </EnvironmentVariable>
-         <EnvironmentVariable
-            key = "TOOLCHAINS"
-            value = "${TOOLCHAINS}"
-            isEnabled = "YES">
-         </EnvironmentVariable>
-      </EnvironmentVariables>
+      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
    <ProfileAction
-      buildConfiguration = "Profile"
+      buildConfiguration = "Release"
       shouldUseLaunchSchemeArgsEnv = "YES"
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "D0E7B63119E9C64500EDBA4D"
-            BuildableName = "carthage.app"
-            BlueprintName = "carthage"
-            ReferencedContainer = "container:Carthage.xcodeproj">
+            BlueprintIdentifier = "D0D1216C19E87B05005E4BAA"
+            BuildableName = "XCDBLD.framework"
+            BlueprintName = "XCDBLD"
+            ReferencedContainer = "container:XCDBLD.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
